refactor: split trade insights ai panel rendering
